Shorter
Also preserves the original features where some threads are spaces and not characters.  The loop forever is accomplished with a never-ending for loop allowing the init line to be combined with the start of the main loop.  Next statements work the same as GOTO but they allow jumping back to the middle of a line. 😄  The INT statements are mostly unnecessary since apple basic treats everything as floating point and it is the functions themselves that strip off the mantissa and use the integer portion.  If you run a value through INT and store it into a variable, it will be converted back to float anyway so it just wastes CPU cycles.  I do use INT in one place to normalize a value as 0 or 1, and that is also how I eliminate an if statement by just multiplying the result by 96 -- it will either be 0 (space) or some other character as a result, etc.